+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Print out future deprecation warning for methods `with_reputation` and `with_normalized_reputation`.
+
+* Fix a finder related bug.
+
 ## ActiveRecordReputationSystem 2.0.0 ##
 
 * Deprecate `init_value` option.